refactor(onboarding): extract demo fallback assets
