+++ src/pollypm/defaults/demo/.gitignore
@@ -1,0 +1,3 @@
+.pollypm/
+__pycache__/
+.pytest_cache/
