+++ src/pollypm/defaults/demo/README.md
@@ -1,0 +1,12 @@
+# PollyPM Demo Repo
+
+This repo is the offline fallback used by onboarding when no recent local
+projects are detected.
+
+## What To Try
+
+- Run `python -m unittest discover -s tests -p "test_*.py"`
+- Read `demo_app.py`
+- Ask Polly to explain, extend, or test the queue summary behavior
+
+Everything in this repo uses only the Python standard library.
